Now we can track the locations of people who are attending user's event
Summary:
Location tracker for guests now works perfectly. I'm in the process of adding the positions of the guests
of a particular event.

Test Plan: Create events and use them to test whether things are working

Reviewers: harshdeep, aykamko, amurthy

Reviewed By: amurthy

Differential Revision: https://phabricator.fb.com/D904489

Now tracks location of everyone attending event user is hosting

Summary:
Location tracker for guests now works perfectly. I'm in the process of adding the positions of the guests
of a particular event.

Test Plan: Create events and use them to test whether things are working

Reviewers: harshdeep, aykamko, amurthy, xianalice

Reviewed By: aykamko

Differential Revision: https://phabricator.fb.com/D909808